+++ docs/prerender.config.ts
@@ -4,7 +4,7 @@ import {PrerenderHydrateOptions} from '@stencil/core/internal/stencil-public-com
 export const config: PrerenderConfig = {
   hydrateOptions(url?: URL): PrerenderHydrateOptions {
     const hydrate: PrerenderHydrateOptions = {
-      excludeComponents: ['deckgo-highlight-code', 'deckgo-slide-poll'],
+      excludeComponents: ['deckgo-highlight-code', 'deckgo-slide-poll', 'deckgo-word-cloud'],
     };
     return hydrate;
   },