+++ docs/src/app/pages/docs/components/app-components-highlight-code/app-components-highlight-code.tsx
@@ -151,8 +151,6 @@ export class AppComponentsHighlightCode {
             <code slot="code">
               &lt;script type=&quot;module&quot;
               src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;highlight-code@latest&#47;dist&#47;deckdeckgo-highlight-code&#47;deckdeckgo-highlight-code.esm.js&quot;&gt;&lt;&#47;script&gt;
-              {'\n'}&lt;script nomodule=&quot;&quot;
-              src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;highlight-code@latest&#47;dist&#47;deckdeckgo-highlight-code&#47;deckdeckgo-highlight-code.js&quot;&gt;&lt;&#47;script&gt;
             </code>
           </deckgo-highlight-code>
           <h3 id="app-components-highlight-code-install-deckdeckgo-highlight-code-from-npm">Install DeckDeckGo Highlight Code from NPM</h3>