+++ docs/src/app/pages/docs/components/app-components-social/app-components-social.tsx
@@ -78,8 +78,6 @@ export class AppComponentsSocial {
             <code slot="code">
               &lt;script type=&quot;module&quot;
               src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;social@latest&#47;dist&#47;deckdeckgo-social&#47;deckdeckgo-social.esm.js&quot;&gt;&lt;&#47;script&gt;
-              {'\n'}&lt;script nomodule=&quot;&quot;
-              src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;social@latest&#47;dist&#47;deckdeckgo-social&#47;deckdeckgo-social.js&quot;&gt;&lt;&#47;script&gt;
             </code>
           </deckgo-highlight-code>
           <h3 id="app-components-social-install-from-npm">Install from NPM</h3>