+++ docs/src/app/pages/docs/slides/app-slide-code/app-slide-code.tsx
@@ -108,8 +108,6 @@ export class AppSlideCode {
             <code slot="code">
               &lt;script type=&quot;module&quot;
               src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;slide-code@latest&#47;dist&#47;deckdeckgo-slide-code&#47;deckdeckgo-slide-code.esm.js&quot;&gt;&lt;&#47;script&gt;
-              {'\n'}&lt;script nomodule=&quot;&quot;
-              src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;slide-code@latest&#47;dist&#47;deckdeckgo-slide-code&#47;deckdeckgo-slide-code.js&quot;&gt;&lt;&#47;script&gt;
             </code>
           </deckgo-highlight-code>
           <h3 id="app-slide-code-from-npm">From NPM</h3>