+++ docs/src/app/pages/docs/slides/app-slide-qrcode/app-slide-qrcode.tsx
@@ -107,8 +107,6 @@ export class AppSlideQRCode {
             <code slot="code">
               &lt;script type=&quot;module&quot;
               src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;slide-qrcode@latest&#47;dist&#47;deckdeckgo-slide-qrcode&#47;deckdeckgo-slide-qrcode.esm.js&quot;&gt;&lt;&#47;script&gt;
-              {'\n'}&lt;script nomodule=&quot;&quot;
-              src=&quot;https:&#47;&#47;unpkg.com&#47;@deckdeckgo&#47;slide-qrcode@latest&#47;dist&#47;deckdeckgo-slide-qrcode&#47;deckdeckgo-slide-qrcode.js&quot;&gt;&lt;&#47;script&gt;
             </code>
           </deckgo-highlight-code>
           <h3 id="app-slide-qrcode-from-npm">From NPM</h3>