+++ webcomponents/core/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 6.1.2 (2020-10-23)
+
+### Fix
+
+- es5 has been dropped with v2, update `unpkg` reference
+
 # 6.1.1 (2020-10-21)
 
 ### Features